+++ lambdas/footyAppStatus/app.py
