+++ lambdas/footyAppStatus/requirements.txt
@@ -1,0 +1,1 @@
+boto3
